[REFACTOR] Reorganize speech and hearing impairment processing
