+++ pandora-common/src/character/index.ts
@@ -4,4 +4,3 @@ export * from './characterTypes';
 export * from './pronouns';
 export * from './restrictionsManager';
 export * from './restrictionTypes';
-export * from './speech';
