+++ pandora-common/src/chat/chatMessageFilter.ts
@@ -1,0 +1,6 @@
+import type { IChatSegment } from './chat';
+
+export interface ChatMessageFilter {
+	isActive(): boolean;
+	processMessage(content: IChatSegment[]): IChatSegment[];
+}
