+++ pandora-common/src/chat/index.ts
@@ -1,2 +1,5 @@
 export * from './chat';
 export * from './chatActions';
+export * from './chatMessageFilter';
+export * from './hearingImpairment';
+export * from './muffling';
